+++ android/src/main/kotlin/com/smileidentity/flutter/SmileIDSmartSelfieEnrollment.kt
@@ -4,8 +4,9 @@ import android.content.Context
 import androidx.compose.runtime.Composable
 import com.smileidentity.SmileID
 import com.smileidentity.compose.SmartSelfieEnrollment
+import com.smileidentity.flutter.results.SmartSelfieCaptureResult
+import com.smileidentity.flutter.utils.SelfieCaptureResultAdapter
 import com.smileidentity.results.SmileIDResult
-import com.smileidentity.util.randomJobId
 import com.smileidentity.util.randomUserId
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.StandardMessageCodec
@@ -28,15 +29,40 @@ internal class SmileIDSmartSelfieEnrollment private constructor(
         val extraPartnerParams = args["extraPartnerParams"] as? Map<String, String> ?: emptyMap()
         SmileID.SmartSelfieEnrollment(
             userId = args["userId"] as? String ?: randomUserId(),
-            jobId = args["jobId"] as? String ?: randomJobId(),
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
             allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
             showInstructions = args["showInstructions"] as? Boolean ?: true,
+            skipApiSubmission = args["skipApiSubmission"] as? Boolean ?: false,
             extraPartnerParams = extraPartnerParams.toImmutableMap(),
         ) {
             when (it) {
-                is SmileIDResult.Success -> onSuccess(it.data)
+                is SmileIDResult.Success -> {
+                    val result =
+                        SmartSelfieCaptureResult(
+                            selfieFile = it.data.selfieFile,
+                            livenessFiles = it.data.livenessFiles,
+                            apiResponse = it.data.apiResponse,
+                        )
+                    val moshi =
+                        SmileID.moshi
+                            .newBuilder()
+                            .add(SelfieCaptureResultAdapter.FACTORY)
+                            .build()
+                    val json =
+                        try {
+                            moshi
+                                .adapter(SmartSelfieCaptureResult::class.java)
+                                .toJson(result)
+                        } catch (e: Exception) {
+                            onError(e)
+                            return@SmartSelfieEnrollment
+                        }
+                    json?.let { js ->
+                        onSuccessJson(js)
+                    }
+                }
+
                 is SmileIDResult.Error -> onError(it.throwable)
             }
         }