+++ android/src/main/kotlin/com/smileidentity/flutter/enhanced/SmileIDSmartSelfieAuthenticationEnhanced.kt
@@ -1,65 +1,64 @@
-package com.smileidentity.flutter
+package com.smileidentity.flutter.enhanced
 
 import android.content.Context
 import androidx.compose.runtime.Composable
 import com.smileidentity.SmileID
-import com.smileidentity.compose.SmartSelfieAuthentication
-import com.smileidentity.flutter.results.SmartSelfieCaptureResult
-import com.smileidentity.flutter.utils.SelfieCaptureResultAdapter
+import com.smileidentity.compose.SmartSelfieAuthenticationEnhanced
+import com.smileidentity.flutter.SmileComposablePlatformView
+import com.smileidentity.networking.FileAdapter
+import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.util.randomUserId
+import com.squareup.moshi.Moshi
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 import kotlinx.collections.immutable.toImmutableMap
 
-internal class SmileIDSmartSelfieAuthentication private constructor(
+internal class SmileIDSmartSelfieAuthenticationEnhanced private constructor(
     context: Context,
     viewId: Int,
     messenger: BinaryMessenger,
     args: Map<String, Any?>,
 ) : SmileComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
     companion object {
-        const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthentication"
+        const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthenticationEnhanced"
     }
 
     @Composable
     override fun Content(args: Map<String, Any?>) {
         val extraPartnerParams = args["extraPartnerParams"] as? Map<String, String> ?: emptyMap()
-        SmileID.SmartSelfieAuthentication(
+        SmileID.SmartSelfieAuthenticationEnhanced(
             userId = args["userId"] as? String ?: randomUserId(),
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
-            allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
             showInstructions = args["showInstructions"] as? Boolean ?: true,
-            skipApiSubmission = args["skipApiSubmission"] as? Boolean ?: false,
             extraPartnerParams = extraPartnerParams.toImmutableMap(),
         ) {
+            val moshi =
+                Moshi.Builder()
+                    .add(FileAdapter)
+                    .build()
             when (it) {
                 is SmileIDResult.Success -> {
                     val result =
-                        SmartSelfieCaptureResult(
+                        SmartSelfieResult(
                             selfieFile = it.data.selfieFile,
                             livenessFiles = it.data.livenessFiles,
                             apiResponse = it.data.apiResponse,
                         )
-                    val moshi =
-                        SmileID.moshi
-                            .newBuilder()
-                            .add(SelfieCaptureResultAdapter.FACTORY)
-                            .build()
                     val json =
                         try {
                             moshi
-                                .adapter(SmartSelfieCaptureResult::class.java)
+                                .adapter(SmartSelfieResult::class.java)
                                 .toJson(result)
                         } catch (e: Exception) {
                             onError(e)
-                            return@SmartSelfieAuthentication
+                            return@SmartSelfieAuthenticationEnhanced
                         }
-                    json?.let { js ->
-                        onSuccessJson(js)
+                    json?.let { response ->
+                        onSuccessJson(response)
                     }
                 }
 
@@ -77,7 +76,7 @@ internal class SmileIDSmartSelfieAuthentication private constructor(
             args: Any?,
         ): PlatformView {
             @Suppress("UNCHECKED_CAST")
-            return SmileIDSmartSelfieAuthentication(
+            return SmileIDSmartSelfieAuthenticationEnhanced(
                 context,
                 viewId,
                 messenger,