+++ lib/smile_id_smart_selfie_enrollment.dart
@@ -21,14 +21,14 @@ class SmileIDSmartSelfieEnrollment extends StatelessWidget {
 
   factory SmileIDSmartSelfieEnrollment({
     Key? key,
-    // userId and jobId can't actually be null in the native SDK but we delegate their creation to
+    // userId can't actually be null in the native SDK but we delegate their creation to
     // the native platform code, since that's where the random ID creation happens
     String? userId,
-    String? jobId,
     bool allowNewEnroll = false,
     bool allowAgentMode = false,
     bool showAttribution = true,
     bool showInstructions = true,
+    bool skipApiSubmission = false,
     Map<String, String>? extraPartnerParams,
     required Function(String resultJson) onSuccess,
     required Function(String errorMessage) onError,
@@ -38,12 +38,12 @@ class SmileIDSmartSelfieEnrollment extends StatelessWidget {
       onError: onError,
       creationParams: {
         "userId": userId,
-        "jobId": jobId,
         "allowNewEnroll": allowNewEnroll,
         "allowAgentMode": allowAgentMode,
         "showAttribution": showAttribution,
         "showInstructions": showInstructions,
         "extraPartnerParams" : extraPartnerParams,
+        "skipApiSubmission" : skipApiSubmission,
       },
     );
   }